Decoding tampered documents now throws SignatureException (due to System.Cryptography limits). Updated test suite.
