+++ Extensions/SignedFileExtensions.cs
@@ -3,10 +3,10 @@ using System.IO;
 using System.Security.Cryptography;
 using System.Security.Cryptography.Pkcs;
 using System.Security.Cryptography.X509Certificates;
-using System.Text.RegularExpressions;
 
 namespace FatturaElettronica.Extensions
 {
+
     public static class SignedFileExtensions
     {
 
@@ -21,7 +21,7 @@ namespace FatturaElettronica.Extensions
         public static void ReadXmlSigned(this FatturaBase fattura, string filePath, bool validateSignature = true)
         {
             using var inputStream = new FileStream(filePath, FileMode.Open, FileAccess.Read);
-            using var contentStream = PreprocessStreamEncoding(inputStream);
+            using var contentStream = StreamExtensions.PreprocessStreamEncoding(inputStream);
             ReadXmlSigned(fattura, contentStream, validateSignature);
 
         }
@@ -30,29 +30,6 @@ namespace FatturaElettronica.Extensions
         {
             ReadXmlSigned(fattura, new MemoryStream(Convert.FromBase64String(File.ReadAllText(filePath))),
                 validateSignature);
-        }
-
-        /// <summary>
-        /// This helper performs preliminary checks to determine if content is likely to be Base64 encoded and if it is it attempts to Decode it and returns.
-        /// </summary>
-        /// <param name="fileContents">The Stream to be processed</param>
-        /// <returns>If the input stream was Base encoded and was successfully decoded it returns a new Stream containing the decoded data. 
-        ///Else it rewinds and returns the original stream</returns>
-        ///<exception cref="FormatException">Although unlikely, if provided with an invalid base encoded Stream that still got through the length check and regex it will throw a FormatException</exception>
-        private static Stream PreprocessStreamEncoding(Stream fileContents)
-        {
-            //Using the full constructor to prevent stream collection
-            using var reader = new StreamReader(fileContents, System.Text.Encoding.UTF8, true, 1024, true);
-            var content = reader.ReadToEnd();
-            //The length check is a bit faster separate instead of inserted in the regex match
-            if (content.Length % 4 != 0 || !Regex.IsMatch(content, "^[A-Za-z0-9+/]*([AQgw]==|[AEIMQUYcgkosw048]=)?$"))
-            {
-                //Unlikely to be B64,we rewind the stream for convenience
-                fileContents.Position = 0;
-                return fileContents;
-            }
-            //Returning a B64Decoded Stream
-            return new MemoryStream(Convert.FromBase64String(content));
         }
 
         public static void ReadXmlSigned(this FatturaBase fattura, Stream stream, bool validateSignature = true)
@@ -84,13 +61,15 @@ namespace FatturaElettronica.Extensions
                     signedFile.CheckSignature(true);
                 }
             }
-            catch (CryptographicException ce) when (validateSignature)
+            catch (CryptographicException ce)//when(validateSignature)
             {
                 throw new SignatureException(Resources.ErrorMessages.SignatureException, ce);
             }
 
             var memoryStream = new MemoryStream();
             memoryStream.Write(signedFile.ContentInfo.Content, 0, signedFile.ContentInfo.Content.Length);
+
+
             return memoryStream;
 
             static byte[] ReadAllBytes(Stream stream)
