+++ FatturaBase.cs
@@ -37,14 +37,15 @@ namespace FatturaElettronica
 
         public static FatturaBase CreateInstanceFromXml(Stream stream, bool validateSignature = false)
         {
+            using var contentStream = StreamExtensions.PreprocessStreamEncoding(stream);
             try
             {
-                return CreateInstanceFromPlainXml(stream);
+                return CreateInstanceFromPlainXml(contentStream);
             }
             catch (XmlException)
             {
                 stream.Position = 0;
-                return CreateInstanceFromXmlSigned(stream, validateSignature);
+                return CreateInstanceFromXmlSigned(contentStream, validateSignature);
             }
         }
