+++ Test/FatturaElettronica.Test.csproj
@@ -39,6 +39,9 @@
     <None Update="Samples\IT02182030391_31.Base64.xml.p7m">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
+    <None Update="Samples\IT02182030391_31_tampered.Base64.xml.p7m">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </None>
     <None Update="Samples\IT02182030391_31_tampered.xml.p7m">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
